refactor: merge eval/ into evaluation/ and remove duplicate model.py
